edited engine objects
