+++ packages/client/src/engine/Core/const.ts
@@ -1,0 +1,22 @@
+export const FPS = 60
+
+export const SECOND = 1000
+
+export const INITIAL_GAME_PARAMS = {
+  startSpeed: 5,
+  maxSpeed: 30,
+  diffSpeed: 2,
+}
+
+export const INITIAL_ENGINE_PROGRESS = {
+  intervalId: null,
+  sessionId: 0,
+  timestamp: 0,
+}
+
+export const INITIAL_PLAYER_PROGRESS = {
+  speed: 0,
+  distance: 0,
+  playTime: 0, // В секундах
+  scores: 0,
+}
