+++ packages/client/src/engine/Objects/Barrier/types.ts
@@ -1,4 +1,4 @@
-import { BaseObjectSpecs } from '../Base/types'
+import { BaseObjectSpecs } from '../Base/_types'
 
 export interface BarrierObjectSpecs extends BaseObjectSpecs {
   image?: CanvasImageSource
