+++ packages/client/src/engine/Objects/Base/index.ts
@@ -1,43 +1,38 @@
 import { canvas } from '@/utils'
-import { BaseObjectSpecs } from './types'
+import CodeBustersEngine from '@/engine/Core'
+import { BaseGameObjectSpecs } from './types'
 
-/*
- * @INFO Базовый объект
- *
- * Служит для описания общих характеристик объектов
- *
- */
+export type BaseGameObjectType = BaseGameObject<BaseGameObjectSpecs>
 
-export default abstract class BaseObject<TObjectSpecs extends BaseObjectSpecs> {
+export default abstract class BaseGameObject<
+  TGameObjectSpecs extends BaseGameObjectSpecs
+> {
+  protected engine: CodeBustersEngine | null = null
   protected canvasApi: ReturnType<typeof canvas>
-  protected deltaTime = 0
-  protected specs: TObjectSpecs | null = null
+  protected specs: TGameObjectSpecs = {} as TGameObjectSpecs
 
-  constructor(canvasApi: ReturnType<typeof canvas>) {
-    this.canvasApi = canvasApi
-  }
+  public key = ''
 
-  protected updateSpecs(
-    newSpecs: Partial<TObjectSpecs>,
-    callback?: (
-      prevSpecs: TObjectSpecs,
-      newSpecs: Partial<TObjectSpecs>
-    ) => TObjectSpecs
+  constructor(
+    key: string,
+    canvasApi: ReturnType<typeof canvas>,
+    initialSpecs: Partial<TGameObjectSpecs> = {}
   ) {
-    if (this.specs && typeof callback === 'function') {
-      callback(this.specs, newSpecs)
+    this.key = key
+    this.canvasApi = canvasApi
 
-      return (this.specs = { ...this.specs, ...newSpecs })
+    this.specs = {
+      ...this.specs,
+      ...(initialSpecs as TGameObjectSpecs),
     }
   }
 
-  public draw(deltaTime: number, specs: TObjectSpecs) {
-    this.deltaTime = deltaTime
-    this.specs = specs
-  }
+  public abstract bindEngine(engine: CodeBustersEngine): void
 
-  public getUniqSpecs(): Partial<TObjectSpecs> | null {
-    return null
+  public draw(specs?: Partial<TGameObjectSpecs>) {
+    if (specs) {
+      this.specs = { ...this.specs, ...specs } as TGameObjectSpecs
+    }
   }
 
   public getSpecs() {
