+++ packages/client/src/engine/Objects/Car/_types.ts
@@ -1,0 +1,10 @@
+import { BaseObjectSpecs } from '@/engine/Objects/Base/types'
+
+export interface CarObjectSpecs extends BaseObjectSpecs {
+  image: CanvasImageSource
+}
+
+export enum CarKeyboadControlEventKey {
+  LEFT = 'ArrowLeft',
+  RIGHT = 'ArrowRight',
+}
