+++ packages/client/src/engine/Objects/Car/const.ts
@@ -1,0 +1,9 @@
+export const INITIAL_SPECS = {
+  x: 0,
+  y: 0,
+  width: 70,
+  height: 140,
+  fill: '#fff',
+  image: new Image(),
+  sensitivity: 40,
+}
