+++ packages/client/src/engine/Objects/Car/index.ts
@@ -1,6 +1,15 @@
-import { BaseObject } from '@/engine/Objects'
-import { CarObjectSpecs, CarKeyboadControlEventKey } from './types'
 import { canvas } from '@/utils'
+import { CodeBustersEngine } from '@/engine'
+import { EngineEvent } from '@/engine/Core/types'
+import { BaseGameObject } from '@/engine/Objects'
+import { INITIAL_SPECS } from './const'
+import { CarObjectSpecs, CarKeyboardControlEventKey } from './types'
+/*
+ * @INFO Объект машины
+ *
+ * Включает в себя описание отрисовки машины и, возможно, других машин на трассе
+ *
+ */
 
 // Получаем координаты машинки
 function getCarPosition() {
@@ -12,54 +21,32 @@ function getCarPosition() {
   }
 }
 
-/*
- * @INFO Объект машины
- *
- * Включает в себя описание отрисовки машины и, возможно, других машин на трассе
- *
- */
-export default class CarObject extends BaseObject<CarObjectSpecs> {
-  static dimensions = {
-    width: 70,
-    height: 140,
-    yAxisPosition: 0,
-    bottomMargin: 30,
-  }
+export default class CarObject extends BaseGameObject<CarObjectSpecs> {
+  protected specs: CarObjectSpecs = INITIAL_SPECS
 
-  static sensitivity = 40
-
-  static createBaseCarSpecs(
-    carImageSrc: string,
-    x: number,
-    y: number,
-    trackHeight: number
+  constructor(
+    key: string,
+    canvasApi: ReturnType<typeof canvas>,
+    initialSpecs: Partial<CarObjectSpecs> = {}
   ) {
-    const carCanvasImage = new Image()
-    carCanvasImage.src = carImageSrc
-    CarObject.dimensions.yAxisPosition =
-      trackHeight -
-      CarObject.dimensions.height -
-      CarObject.dimensions.bottomMargin
+    super(key, canvasApi, initialSpecs)
 
-    return {
-      image: carCanvasImage,
-      x,
-      y,
-      width: CarObject.dimensions.width,
-      height: CarObject.dimensions.height,
+    if (initialSpecs.image) {
+      this.specs.image = initialSpecs.image
     }
   }
 
-  constructor(protected canvasApi: ReturnType<typeof canvas>) {
-    super(canvasApi)
+  public bindEngine(engine: CodeBustersEngine) {
+    this.engine = engine
 
-    this.onKeyDown = this.onKeyDown.bind(this)
+    this.engine.eventEmitter.on(
+      EngineEvent.PRESS_KEY,
+      this.onPressKey.bind(this)
+    )
   }
 
   public drawCar() {
     if (this.canvasApi.ctx && this.specs) {
-      this.clear()
-
       const carPosition = getCarPosition()
 
       this.canvasApi.ctx.beginPath()
@@ -78,48 +65,37 @@ export default class CarObject extends BaseObject<CarObjectSpecs> {
     }
   }
 
-  public draw(delta: number, specs: CarObjectSpecs) {
-    this.specs = specs
+  public draw(specs?: Partial<CarObjectSpecs>) {
+    if (specs) {
+      this.specs = { ...this.specs, ...specs }
+    }
 
-    this.drawCar()
+    if (this.canvasApi.ctx) {
+      this.clear()
+      this.drawCar()
+    }
   }
 
-  public getCenterOnTrack(trackWidth: number) {
-    return trackWidth / 2 - CarObject.dimensions.width / 2
-  }
-
-  private onKeyDown(event: KeyboardEvent) {
+  private onPressKey(event: KeyboardEvent) {
     const prevSpecs = this.specs as CarObjectSpecs
 
     switch (event.key) {
-      case CarKeyboadControlEventKey.LEFT:
-        this.clear()
-
-        this.draw(0, {
-          ...prevSpecs,
-          x: prevSpecs.x - CarObject.sensitivity,
+      case CarKeyboardControlEventKey.LEFT:
+        this.draw({
+          x:
+            prevSpecs.x - (this.specs.sensitivity || INITIAL_SPECS.sensitivity),
         })
         break
 
-      case CarKeyboadControlEventKey.RIGHT:
-        this.clear()
-
-        this.draw(0, {
-          ...prevSpecs,
-          x: prevSpecs.x + CarObject.sensitivity,
+      case CarKeyboardControlEventKey.RIGHT:
+        this.draw({
+          x:
+            prevSpecs.x + (this.specs.sensitivity || INITIAL_SPECS.sensitivity),
         })
         break
 
       default:
         break
     }
-  }
-
-  public addListeners() {
-    document.addEventListener('keydown', this.onKeyDown)
-  }
-
-  public removeListeners() {
-    document.removeEventListener('keydown', this.onKeyDown)
   }
 }
