+++ packages/client/src/engine/Objects/Line/const.ts
@@ -1,0 +1,7 @@
+export const INITIAL_SPECS = {
+  x: 0,
+  y: 0,
+  width: 10,
+  height: 140,
+  fill: '#fff',
+}
