+++ packages/client/src/engine/Objects/Line/types.ts
@@ -1,3 +1,3 @@
 import { BaseGameObjectSpecs } from '@/engine/Objects/Base/types'
 
-export type TrackObjectSpecs = BaseGameObjectSpecs
+export type LineObjectSpecs = BaseGameObjectSpecs
