+++ packages/client/src/engine/Objects/Track/const.ts
@@ -1,0 +1,7 @@
+export const INITIAL_SPECS = {
+  x: 0,
+  y: 0,
+  width: 500,
+  height: 0,
+  fill: 'gray',
+}
