+++ packages/client/src/engine/Objects/index.ts
@@ -1,3 +1,4 @@
-export { default as BaseObject } from './Base'
+export { default as BaseGameObject } from './Base'
 export { default as TrackObject } from './Track'
 export { default as CarObject } from './Car'
+export { default as LineObject } from './Line'
