+++ packages/client/src/pages/Play/index.tsx
@@ -1,21 +1,25 @@
-import { RunMethodOptions } from '@/engine/Core/types'
+import { useRef, useState, useMemo, useEffect } from 'react'
+import { EngineStartMethodOptions } from '@/engine/Core/types'
 import { CarObject, TrackObject } from '@/engine/Objects'
 import { MainLayout } from '@/layouts'
 import { selectGameScores } from '@/store/slices/gameSlice/selectrors'
 import { setLeaderboardData } from '@/store/slices/leaderboardSlice/thunks'
 import { selectUserInfo } from '@/store/slices/userSlice/selectors'
 import { useAppDispatch, useAppSelector } from '@/store/typedHooks'
-import classNames from 'classnames/bind'
-import { useRef, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { GameControls, PlayerScores } from './components'
 import { useEngine, useMakeFullscreen, useCountry } from './hooks'
+import { CircularProgress } from '@mui/material'
+
+import classNames from 'classnames/bind'
 import styles from './styles.module.scss'
 
 const cx = classNames.bind(styles)
 
 const PlayPage = () => {
   const [level, setLevel] = useState(1)
+  const [isLoading, setLoading] = useState(true)
+
   const user = useAppSelector(selectUserInfo)
   const scores = useAppSelector(selectGameScores)
   const dispatch = useAppDispatch()
@@ -38,11 +42,15 @@ const PlayPage = () => {
     endGameMessageRef,
   })
 
+  const trackSpecs = engine?.getGameObject<TrackObject>('track').getSpecs()
+
+  const carSpecs = engine?.getGameObject<CarObject>('car').getSpecs()
+
   useMakeFullscreen()
 
   const country = useCountry()
 
-  const startGame = (options?: RunMethodOptions) => {
+  const startGame = (options?: EngineStartMethodOptions) => {
     setLevel(1)
 
     engine?.start(options)
@@ -83,6 +91,12 @@ const PlayPage = () => {
     }
   }
 
+  useEffect(() => {
+    if (trackSpecs && carSpecs) {
+      setLoading(false)
+    }
+  }, [trackSpecs, carSpecs])
+
   return (
     <MainLayout>
       <div className={cx('play__page')}>
@@ -96,7 +110,14 @@ const PlayPage = () => {
             controls={{ startGame, pauseGame, endGame, leaderboard, exitGame }}
           />
         </div>
-        <div ref={containerRef} className={cx('play__area')}>
+        {isLoading && (
+          <div className={cx(['play__loader'])}>
+            <CircularProgress />
+          </div>
+        )}
+        <div
+          ref={containerRef}
+          className={cx('play__area', { ['play__area_hidden']: isLoading })}>
           <canvas
             ref={backgroundRef}
             className={cx('play__background')}
@@ -107,20 +128,20 @@ const PlayPage = () => {
           <canvas
             ref={carRef}
             className={cx('play__car')}
-            width={TrackObject.width}
-            height={CarObject.dimensions.height}
-            style={{ bottom: CarObject.dimensions.bottomMargin }}
+            width={trackSpecs?.width || 0}
+            height={carSpecs?.height || 0}
+            style={{ bottom: 30 }}
           />
           <canvas
             ref={barrierRef}
             className={cx('play__barrier')}
-            width={TrackObject.width}
-            height={trackRef?.current?.height || 710}
+            width={trackSpecs?.width || 0}
+            height={trackSpecs?.height || 710}
           />
           <canvas
             ref={endGameMessageRef}
             className={cx('play__endgame')}
-            width={TrackObject.width}
+            width={trackSpecs?.width || 0}
             height={trackRef?.current?.height || 710}
           />
         </div>
